Moved test files to new "sandbox" directory, for practicing various React implementations.
