+++ frontend/src/sandbox/test.tsx
@@ -8,7 +8,7 @@
 
 import { useState } from 'react';
 import axios from 'axios';
-import Test_searchbar from '../components/test_searchbar.tsx';
+import Test_searchbar from '../sandbox/test_searchbar.tsx';
 
 export default function Test() {
